40415: Failed test: Bestehende Umfrage importieren

+++ components/ILIAS/Survey/classes/class.ilObjSurvey.php
@@ -321,7 +321,9 @@ class ilObjSurvey extends ilObject
         );
         $this->deleteAllUserData(false);
 
-        $this->code_manager->deleteAll(true);
+        if (isset($this->code_manager)) {
+            $this->code_manager->deleteAll(true);
+        }
 
         // delete export files
         $svy_data_dir = ilFileUtils::getDataDir() . "/svy_data";
